Version 1.3 Update

+++ src/QIFLibrary.Tests/IO/CSVWriterTests.cs
@@ -12,17 +12,20 @@ public class CSVWriterTests
     public void RoundTripTest()
     {
         var filename = Path.Combine("TestAssets", "sample1.csv");
-        var doc = CSVDocument.ParseFile(filename);
-        var ms = new MemoryStream();
-        var writer = new StreamWriter(ms);
+        if (File.Exists(filename))
+        {
+            var doc = CSVDocument.ParseFile(filename);
+            var ms = new MemoryStream();
+            var writer = new StreamWriter(ms);
 
-        var target = new CSVWriter(writer);
-        target.Write(doc);
-        writer.Flush();
+            var target = new CSVWriter(writer);
+            target.Write(doc);
+            writer.Flush();
 
-        var actual = Encoding.UTF8.GetString(ms.ToArray()).Replace("\n", string.Empty).Replace("\r", string.Empty);
-        var expected = File.ReadAllText(filename).Replace("\"", string.Empty).Replace("\n", string.Empty).Replace("\r", string.Empty);
-        Assert.AreEqual(expected, actual);
+            var actual = Encoding.UTF8.GetString(ms.ToArray()).Replace("\n", string.Empty).Replace("\r", string.Empty);
+            var expected = File.ReadAllText(filename).Replace("\"", string.Empty).Replace("\n", string.Empty).Replace("\r", string.Empty);
+            Assert.AreEqual(expected, actual);
+        }
     }
 
     [TestMethod]
